Update Project Ownerships
