+++ sentry/project_ownerships_test.go
@@ -1,6 +1,7 @@
 package sentry
 
 import (
+	"context"
 	"fmt"
 	"net/http"
 	"testing"
@@ -8,7 +9,7 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestProjectOwnershipService_Get(t *testing.T) {
+func TestProjectOwnershipsService_Get(t *testing.T) {
 	client, mux, _, teardown := setup()
 	defer teardown()
 
@@ -26,7 +27,8 @@ func TestProjectOwnershipService_Get(t *testing.T) {
 		}`)
 	})
 
-	ownership, _, err := client.Ownership.Get("the-interstellar-jurisdiction", "powerful-abolitionist")
+	ctx := context.Background()
+	ownership, _, err := client.ProjectOwnerships.Get(ctx, "the-interstellar-jurisdiction", "powerful-abolitionist")
 	assert.NoError(t, err)
 
 	expected := &ProjectOwnership{
@@ -42,7 +44,7 @@ func TestProjectOwnershipService_Get(t *testing.T) {
 	assert.Equal(t, expected, ownership)
 }
 
-func TestProjectOwnershipService_Update(t *testing.T) {
+func TestProjectOwnershipsService_Update(t *testing.T) {
 	client, mux, _, teardown := setup()
 	defer teardown()
 
@@ -66,7 +68,8 @@ func TestProjectOwnershipService_Update(t *testing.T) {
 	params := &UpdateProjectOwnershipParams{
 		Raw: "# assign issues to the product team, no matter the area\nurl:https://example.com/areas/*/*/products/* #product-team",
 	}
-	ownership, _, err := client.Ownership.Update("the-interstellar-jurisdiction", "the-obese-philosophers", params)
+	ctx := context.Background()
+	ownership, _, err := client.ProjectOwnerships.Update(ctx, "the-interstellar-jurisdiction", "the-obese-philosophers", params)
 	assert.NoError(t, err)
 	expected := &ProjectOwnership{
 		Raw:                "# assign issues to the product team, no matter the area\nurl:https://example.com/areas/*/*/products/* #product-team",
